+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Fixes `getAsyncId` error on Node v8 when using an `Http.Agent` with the `keepAlive` option.
+
 ## 1.3.2 (Nov 10, 2016)
 - Adds compatibility with Node v7.  
   Thanks, [Eric Hacke](https://github.com/ehacke), for the help!